api/pyproject.toml
[project]
name = "haremovie_api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "neka-nat", email = "nekanat.stock@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "fastapi[all]>=0.116.1",
    "google-cloud-aiplatform[adk,agent-engines]>=1.114.0",
    "loguru>=0.7.3",
    "sqlmodel>=0.0.25",
]

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"
